+++ src/views/toggle-button.js
@@ -1,10 +1,7 @@
-import { container as containerDimensions } from '../helpers/dimensions';
-import { createElement, swapElementStyles } from '../helpers/utils';
-import onStateSVGDataString from '../assets/switch-to-qr-icon.svg';
-import offStateSVGDataString from '../assets/switch-to-pdf417-icon.svg';
+import { createElement } from '../helpers/utils';
 
-const BUTTON_MIN_SIZE = 24;
-const OFF_STATE_TIMEOUT = 10;
+const OFF_STATE_TIMEOUT = 10000;
+const DEBOUNCE_TIMEOUT = 350;
 
 /**
  * Class representing a button switching between two states. Exposes a callback
@@ -12,15 +9,13 @@ const OFF_STATE_TIMEOUT = 10;
  *
  * @param {Object} options
  * @param {Number} optinos.w - Width of the container, button will be rendered in.
+ * @param {Number} optinos.h - Height of the container, button will be rendered in.
  * @param {Function} options.onToggle - Callback when button state is toggled.
  */
 export class ToggleButton {
-    constructor({ w, onToggle }) {
-        const stateToggleButtonSize = w * (BUTTON_MIN_SIZE / containerDimensions.minWidth);
-        const svgDataStrings = [onStateSVGDataString, offStateSVGDataString];
-        const [onStateImageElement, offStateImageElement] = ToggleButton.createButtonStateImageElements(svgDataStrings, stateToggleButtonSize);
-        this._buttonTouchAreaElement = ToggleButton.createTouchAreaButton(stateToggleButtonSize, [onStateImageElement, offStateImageElement]);
-        this._buttonTouchAreaElement.addEventListener('click', ToggleButton.createToggleEventListener(onStateImageElement, offStateImageElement, onToggle));
+    constructor({ w, h, onToggle }) {
+        this._buttonTouchAreaElement = ToggleButton.createTouchAreaButton(w, h);
+        this._buttonTouchAreaElement.addEventListener('click', ToggleButton.createToggleEventListener(onToggle));
     }
 
     /**
@@ -35,87 +30,53 @@ export class ToggleButton {
     /**
      * Create the actual button element.
      *
-     * @param {Number} size - The size of the square button.
-     * @param {Array.<HTMLElement>} children - The child elements
+     * @param {Number} w - The width of the button.
+     * @param {Number} h - The height of the button.
      *
      * @returns {HTMLDivElement}
      */
-    static createTouchAreaButton(size, children) {
+    static createTouchAreaButton(w, h) {
         return createElement('div', null,
             {
                 position: 'absolute',
-                width: `${size}px`,
-                height: `${size}px`,
+                width: `${w}px`,
+                height: `${h}px`,
                 right: '0px',
                 bottom: `0px`
-            },
-            children
+            }
         );
     };
 
     /**
      * Create the event listener to toggle button states.
      *
-     * @param {HTMLImageElement} onElement - Element to show when in on state.
-     * @param {HTMLImageElement} offElement - Element to show when in off state.
      * @param {Function} [toggleCallback] - Callback for when animation is complete.
      *
      * @returns {EventListener}
      */
-    static createToggleEventListener(onElement, offElement, toggleCallback, offStateTimeout = OFF_STATE_TIMEOUT) {
+    static createToggleEventListener(toggleCallback, debounceTimeout = DEBOUNCE_TIMEOUT, offStateTimeout = OFF_STATE_TIMEOUT) {
         let isOnState = true;
-        let isAnimating = false;
-        let timeoutId;
+        let isToggling = false;
+        let offStateTimeoutId;
 
         const performToggle = () => {
-            isAnimating = true;
-            // Call `toggleCallback` first, so it can start it's animations at same time.
+            isToggling = true;
+            setTimeout(() => { isToggling = false; }, debounceTimeout);
             typeof toggleCallback === 'function' && toggleCallback(isOnState);
-            swapElementStyles(onElement, offElement, ['opacity'], () => { isAnimating = false; });
         };
 
         return () => {
-            if (isAnimating) { return; }
-            if (timeoutId) { clearTimeout(timeoutId); }
+            if (isToggling) { return; }
+            if (offStateTimeoutId) { clearTimeout(offStateTimeoutId); }
 
             isOnState = !isOnState;
             if (!isOnState) {
-                timeoutId = setTimeout(() => {
+                offStateTimeoutId = setTimeout(() => {
                     isOnState = true;
                     performToggle();
-                }, offStateTimeout * 1000);
+                }, offStateTimeout);
             }
             performToggle();
         };
-    };
-
-    /**
-     * Provided an array SVG files as strings, create and return array of `HTMLImageElement`s for each
-     * file's data.
-     *
-     * @param {Array.<String>} svgDataStrings - Array of SVG file content to convert to `HTMLImageElement`s.
-     * @param {Number} imageSize - The images size
-     *
-     * @returns {Array.<HTMLImageElement>}
-     */
-    static createButtonStateImageElements(svgDataStrings, imageSize, minImageSize = BUTTON_MIN_SIZE) {
-        return svgDataStrings.map((svgDataString, index) => {
-            return createElement('img',
-                {
-                    src: `data:image/svg+xml;base64,${global.btoa(svgDataString)}`,
-                    width: minImageSize,
-                    height: minImageSize
-                },
-                {
-                    position: 'absolute',
-                    // Scale to fit desired `imageSize`.
-                    transform: `scale(${imageSize / minImageSize})`,
-                    transformOrigin: 'top left',
-                    left: '0px',
-                    top: '0px',
-                    opacity: index === 0 ? 100 : 0
-                }
-            );
-        });
     };
 }